交易状态的凭证上传和打回完成，文件上传修改为oss

+++ Refund/RefundO-root/src/main/java/com/refund/root/mapper/RefundRequestsMapper.java
@@ -68,5 +68,5 @@ public interface RefundRequestsMapper
      * @param requestIds 退款申请主键
      * @param status 状态
      * */
-    int updateRefundRequestsStatus(@Param("requestIds") Long[] requestIds, @Param("status") Integer status,@Param("rejectReason") String rejectReason);
+    int updateRefundRequestsStatus(@Param("requestIds") Long[] requestIds, @Param("status") Long status,@Param("rejectReason") String rejectReason);
 }
